+++ backend/canisters/local_user_index/impl/src/model/mod.rs
@@ -5,6 +5,7 @@ pub mod group_event_batch;
 pub mod local_community_map;
 pub mod local_group_map;
 pub mod local_user_map;
+pub mod media_scan_job_log;
 pub mod moderation_queue;
 pub mod premium_items;
 pub mod referral_codes;